requirements.txt
django-allauth
PyJWT
cryptography
